feat(project): add initial implementation of project management crate
This crate is responsible for managing projects in the application.
Projects are the primary organizational structure, consisting of
documents.
This crate also includes saving/loading and undo/redo functionallity.
To implement behaviour for a document in a project, the Module trait can
be used. Examples for future Modules are 'Sketch' or 'Assembly'.