feat(occara): add crate
Occara is a high level binding rust library for OpenCASCADE.
This commit adds the initial framework for writing bindings.

The crate will be written in both C++ and Rust.
Instead of writing bindings for all used OpenCASCADE functions and then implementing
high level bindings in rust, we will instead write the abstraction
inside C++.
This has a couple of advantages:
- We can (in future) leverage autocxx, to generate bindings from C++ to
  Rust. This is not directly possible with OpenCASCADE, since its
  'Handle' leverages templates, that are currently not supported by
  autocxx.
- We can use the highly optimized 'Standard' libraries from OpenCASCADE,
  implementing usefull data structures not found in the Rust standard
  library.
- Lower overhead, since we directly call OpenCASCADE, without any
  bindings inbetween.

Currently this create uses manually written cxx bindings. This can be
changed to autocxx in a later commit.