feat(occara): add basic mesh generation and obj export
This commit adds the ability to generate a mesh representation of a
Shape object, allowing access to its geometry and export as an OBJ file.

This current implementation is however very inefficient due to the use
of automatically generated bindings through autocxx. Manually writing
bindings would accelerate this by a very large margin.