feat(computegraph): add computegraph crate
Add a new crate called `computegraph` which provides functionality to
dynamically build and execute directed acyclic graphs of computations.
This will be later required by the viewport to increase efficiency and
allow workbenches to change the rendering pipeline (like displaying
constraints in a sketch or a handle for changing a extrusion).