feat(project): make project management thread safe
This commit replaces uses of Rc and RefCell to Arc and Mutex
respectively. This is by far not the best way to achieve thread safety,
since acquiring the lock could be done on a per data section basis
instead of a per project one.

This will be addressed with the rewrite to a CRDT based approach for
storing state. This will also remove the countless uses of unwrap().