feat(viewport): associate a Viewport with a ProjectSession
Now each Viewport is coupled to a ProjectSession, which is automatically
passed to all ViewportPlugins and nodes of the SceneGraph.

For now a full ProjectSession is passed, which will be changed in the
future to a read only ProjectSession.

At a later date this mechanism can be optimized by implicitly recording
all accesses to the project, only rerunning the nodes/plugins if the
output could change.